guidelines, homepage and feedback UI completed

+++ app/src/main/java/com/example/mymargdarshaka/MyStudents.java
@@ -1,18 +1,33 @@
 package com.example.mymargdarshaka;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.view.Gravity;
+import android.view.LayoutInflater;
+import android.view.MenuItem;
+import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
+import android.widget.LinearLayout;
+import android.widget.PopupWindow;
+import android.widget.Toast;
+
+import com.google.android.material.navigation.NavigationView;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
 
 public class MyStudents extends AppCompatActivity {
-
-    Button logout;
-
     SharedPreferences sharedPreferences;
+    FirebaseDatabase firebaseDatabase;
+    DatabaseReference databaseReference;
+
+    com.google.android.material.appbar.MaterialToolbar topAppBar;
+    androidx.drawerlayout.widget.DrawerLayout drawerLayout;
+    NavigationView navigationView;
 
     private static final String SHARED_PREF_NAME = "login";
 
@@ -21,22 +36,75 @@ public class MyStudents extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_my_students);
 
-        logout=(Button) findViewById(R.id.logout_student);
-
         sharedPreferences = getSharedPreferences(SHARED_PREF_NAME,MODE_PRIVATE);
 
-        logout.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                SharedPreferences.Editor editor = sharedPreferences.edit();
-                editor.clear();
-                editor.commit();
+        topAppBar = findViewById(R.id.topAppBar);
+        drawerLayout = findViewById(R.id.drawerLayout);
+        navigationView = findViewById(R.id.navigationView);
 
-                Intent i = new Intent(MyStudents.this,MainActivity.class);
-                i.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
-                startActivity(i);
+        topAppBar.setNavigationOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                drawerLayout.openDrawer(Gravity.LEFT);
             }
         });
+
+        navigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+                Toast.makeText(getApplicationContext(), "menu item selected "+item.toString(), Toast.LENGTH_SHORT).show();
+
+                String choice = item.toString();
+                if(choice.equals("Guidelines"))
+                {
+                    Intent i = new Intent(MyStudents.this,GuidelinesForMentors.class);
+                    startActivity(i);
+                }
+                else if(choice.equals("My Students"))
+                {
+                    //code to shift to Student Details Page
+                }
+                else if(choice.equals("Feedback"))
+                {
+                    //code for Feedback
+                    LayoutInflater inflater = (LayoutInflater)
+                            getSystemService(LAYOUT_INFLATER_SERVICE);
+                    View popupView = inflater.inflate(R.layout.feedback_popup, null);
+
+                    // create the popup window
+                    int width = LinearLayout.LayoutParams.WRAP_CONTENT;
+                    int height = LinearLayout.LayoutParams.WRAP_CONTENT;
+                    boolean focusable = true; // lets taps outside the popup also dismiss it
+                    final PopupWindow popupWindow = new PopupWindow(popupView, width, height, focusable);
+
+                    // show the popup window, which view you pass in doesn't matter, it is only used for the window token
+                    View view = findViewById(android.R.id.content).getRootView();
+                    popupWindow.showAtLocation(view, Gravity.CENTER, 0, 0);
+                    // dismiss the popup window when touched
+                    popupView.setOnTouchListener(new View.OnTouchListener() {
+                        @Override
+                        public boolean onTouch(View v, MotionEvent event) {
+                            popupWindow.dismiss();
+                            return true;
+                        }
+                    });
+                }
+                else if(choice.equals("Logout"))
+                {
+                    SharedPreferences.Editor editor = sharedPreferences.edit();
+                    editor.clear();
+                    editor.commit();
+
+                    Intent i = new Intent(MyStudents.this,MainActivity.class);
+                    i.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
+                    startActivity(i);
+                }
+
+                drawerLayout.closeDrawer(Gravity.START, true);
+                return true;
+            }
+        });
+
 
     }
 }
